refactor(config-ui): move connection-status to plugin/component from component
