+++ config-ui/src/plugins/components/index.ts
@@ -18,6 +18,7 @@
 
 export * from './connection-form';
 export * from './connection-list';
+export * from './connection-status';
 export * from './data-scope';
 export * from './data-scope-form';
 export * from './data-scope-form-2';
